requirements.txt
django-taggit==0.9.3